[fix]more페이지 코드 파일 정리
