+++ types/more/notice.ts
@@ -1,0 +1,10 @@
+export type NoticeType = 'normal' | 'update' | 'event' | 'policy';
+
+export interface Notice {
+   id: number;
+   type: NoticeType;
+   title: string;
+   detail: string;
+   date: string;
+   isTopFixed?: boolean;
+}
